+++ test/unit/utils/utils.test.js
@@ -395,10 +395,11 @@ describe('utils', () => {
   });
 
   describe('getTimezoneForDataProcessing', () => {
+    // temporary change until we have consistent upload data
     const data = [
-      { type: 'upload', time: '2018-01-02T00:00:00.000Z', timezone: 'US/Eastern' },
+      { type: 'undefined', time: '2018-01-02T00:00:00.000Z', timezone: 'US/Eastern' },
       { type: 'smbg', time: '2018-01-10T00:00:00.000Z', timezone: 'US/Central' },
-      { type: 'upload', time: '2018-02-02T00:00:00.000Z', timezone: 'US/Pacific' },
+      { type: 'undefined', time: '2018-02-02T00:00:00.000Z', timezone: 'US/Pacific' },
     ];
 
     const queryParams = {};
@@ -455,7 +456,7 @@ describe('utils', () => {
       it('should fall back to browser time when given an invalid timezone', () => {
         const dataWithInvalidTimezone = data.slice();
         dataWithInvalidTimezone.push({
-          type: 'upload',
+          type: 'undefined',
           time: '2018-02-10T00:00:00.000Z',
           timezone: 'invalid',
         });
@@ -477,7 +478,7 @@ describe('utils', () => {
       it('should fall back to timezone-naive display time when given an invalid timezone and cannot determine timezone from browser', () => {
         const dataWithInvalidTimezone = data.slice();
         dataWithInvalidTimezone.push({
-          type: 'upload',
+          type: 'undefined',
           time: '2018-02-10T00:00:00.000Z',
           timezone: 'invalid',
         });
@@ -494,7 +495,7 @@ describe('utils', () => {
       });
     });
 
-    context('Timezone not provided from query params or most recent upload', () => {
+    context('Timezone not provided from query params or most recent upload/undefined', () => {
       it('should fall back to browser timezone if available', () => {
         const DateTimeFormatStub = sinon.stub(Intl, 'DateTimeFormat').returns({
           resolvedOptions: () => {