+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It is based on Tidepool Blip 1.27.
 ## Unreleased
 ### Fixed
 - PT-338 Fix some translations.
+- PT-1115 Fix wrong display due to inconsitent timezone in upload object
 - PT-1108 PT-1105 PT-1114 Make Diabeloop devices automated devices.
 
 ### Changed