+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
-## Unreleased
+## 0.12.0 - 2020-02-17
 ### Fixed
 - PT-338 Fix some translations.
 - PT-1115 Fix wrong display due to inconsitent timezone in upload object